+++ src/commands/install.ts
@@ -1,13 +1,18 @@
 import inquirer from 'inquirer'
 import console from 'consola'
-import { testApi } from '../core/api'
+import { initApi, testApi } from '../core/api'
 import { config } from '../core/config'
 import { validURL } from '../core/utils'
 import { bold } from 'kleur'
 import { installMessage } from '../modules/install'
 
+interface Answers {
+  host: string
+  token: string
+}
+
 export default async function install (): Promise<void> {
-  const result = await inquirer.prompt([
+  const result = await inquirer.prompt<Answers>([
     {
       name    : 'host',
       type    : 'input',
@@ -19,8 +24,8 @@ export default async function install (): Promise<void> {
       name    : 'token',
       type    : 'input',
       message : ({ host }) => `\r\n${installMessage(host)}\r\nInput your token here:`,
-      validate: async (token, { host }) => {
-        const check = await testApi(host, token)
+      validate: async (token: string, answers: Answers) => {
+        const check = await testApi(answers.host, token)
 
         if (!check)
           return 'Invalid personal token'
@@ -32,6 +37,8 @@ export default async function install (): Promise<void> {
 
   config.set('gitlab.host', result.host)
   config.set('gitlab.token', result.token)
+
+  await initApi()
 
   console.success(bold('Install success'))
 }